refactor: #221 - 검색 결과 페이지 client component 분리
