+++ src/components/index.ts
@@ -19,3 +19,4 @@ export { default as Comment } from './common/Comment/Comment'
 export { default as Notification } from './common/Notification/Notification'
 export { default as SpaceList } from './SpaceList/SpaceList'
 export { default as Spinner } from './common/Spinner/Spinner'
+export { default as SearchController } from './SearchController/SearchController'
